mcp_internet_search/requirements.txt
# MCP Internet Search Server Dependencies

# Serper API
requests>=2.31.0

# Voyage AI (for reranking)
voyageai>=0.2.0

# Web scraping
trafilatura>=1.11.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# MCP Server
mcp>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
pydantic>=2.5.0
python-dotenv
